SM-55: move REST API v1 code to its own namespace
