+++ app/API/V1/Http/Resources/ReparationRequestCollection.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Resources;
+namespace App\API\V1\Http\Resources;
 
 use Illuminate\Http\Resources\Json\ResourceCollection;
 
-class ReparationRequestMaterialCollection extends ResourceCollection
+class ReparationRequestCollection extends ResourceCollection
 {
     /**
      * Transform the resource collection into an array.
