+++ app/API/V1/Http/Resources/ReparationRequestMaterialCollection.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Resources;
+namespace App\API\V1\Http\Resources;
 
 use Illuminate\Http\Resources\Json\ResourceCollection;
 
-class ReparationRequestStatusCollection extends ResourceCollection
+class ReparationRequestMaterialCollection extends ResourceCollection
 {
     /**
      * Transform the resource collection into an array.
