+++ app/API/V1/Http/Resources/ReparationRequestResource.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Resources;
+namespace App\API\V1\Http\Resources;
 
 use Illuminate\Http\Resources\Json\JsonResource;
